Update main.cpp
Turned movement in a method of board, to shorten the length of code. Don't know why I didn't do this sooner. I won't be able to do this for eating, because there are too many conditions involved for it to make sense.
Fixed corner and edge logic to work more or less flawlessly. No more out of bound errors or infinite loops.